requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
delta-spark==2.4.0